requirements.txt
asgiref==3.7.2
captcha==0.5.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
crispy-bootstrap4==2023.1
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.57.0
django-crispy-forms==2.1
django-jazzmin==2.6.0
django-recaptcha==3.0.0
django-tinymce==3.6.1
fontawesomefree==6.4.2
idna==3.4
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
